backend/utils/image_processing.py
import cv2
import numpy as np

def process_image_for_analysis(image_path):
    # This would contain actual image preprocessing logic
    img = cv2.imread(image_path)
    if img is None:
        return None
    # Example: Resize to standard size for CNN
    img = cv2.resize(img, (224, 224))
    return img
